Cleaned up single-consumer filter helper exports
ref https://linear.app/ghost/issue/BER-3414

Moved canonicalizeClauses, normalizeMultiValue, getDayBoundsInUtc, and
createOperatorOptions into their single consumer modules as unexported
functions. Removed dead code normalizeOperator and deleted the
filter-operator-options module. filter-normalization.ts now only exports
escapeNqlString, which is the only genuinely shared utility.
