--- apps/posts/src/views/filters/filter-operator-options.ts
@@ -1,20 +1,0 @@
-export interface OperatorOption {
-    value: string;
-    label: string;
-}
-
-interface CreateOperatorOptionsOptions {
-    labels?: Record<string, string>;
-}
-
-export function createOperatorOptions(
-    operators: readonly string[],
-    options: CreateOperatorOptionsOptions = {}
-): OperatorOption[] {
-    const labels = options.labels || {};
-
-    return operators.map(operator => ({
-        value: operator,
-        label: labels[operator] ?? operator.replaceAll('-', ' ')
-    }));
-}
